requirements.txt
opencv-python>=4.8.0
numpy>=1.21.0
ultralytics>=8.0.0
pillow>=9.0.0
matplotlib>=3.5.0
pyyaml>=6.0
playsound>=1.2.2
